+++ InfluxDB.Net.sln
@@ -5,6 +5,7 @@ VisualStudioVersion = 14.0.23107.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{91B2CB56-7AB0-4AF1-A081-95525F8A754C}"
 	ProjectSection(SolutionItems) = preProject
+		InfluxDB.Net\Infrastructure\InfluxDbClient.cs = InfluxDB.Net\Infrastructure\InfluxDbClient.cs
 		.nuget\NuGet.Config = .nuget\NuGet.Config
 		.nuget\NuGet.exe = .nuget\NuGet.exe
 		.nuget\NuGet.targets = .nuget\NuGet.targets
