+++ InfluxDB.Net/Contracts/IInfluxDb.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using InfluxDB.Net.Infrastructure.Influx;
 using InfluxDB.Net.Models;
 
 namespace InfluxDB.Net.Contracts
