+++ InfluxDB.Net/Contracts/IInfluxDbClient.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
+using InfluxDB.Net.Infrastructure.Influx;
 using InfluxDB.Net.Models;
 
 namespace InfluxDB.Net.Contracts
