+++ InfluxDB.Net/InfluxDB.Net.csproj
@@ -44,23 +44,23 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Check.cs" />
-    <Compile Include="DatabaseConfiguration.cs" />
-    <Compile Include="IInfluxDb.cs" />
-    <Compile Include="IInfluxDbClient.cs" />
+    <Compile Include="Infrastructure\Validation\Check.cs" />
+    <Compile Include="Infrastructure\Configuration\DatabaseConfiguration.cs" />
+    <Compile Include="Contracts\IInfluxDb.cs" />
+    <Compile Include="Contracts\IInfluxDbClient.cs" />
     <Compile Include="InfluxDbClient.cs" />
-    <Compile Include="InfluxDbClientConfiguration.cs" />
-    <Compile Include="InfluxDbException.cs" />
-    <Compile Include="InfluxDbResponse.cs" />
-    <Compile Include="IRequestContent.cs" />
-    <Compile Include="JsonRequestContent.cs" />
-    <Compile Include="LogLevel.cs" />
+    <Compile Include="Infrastructure\Configuration\InfluxDbClientConfiguration.cs" />
+    <Compile Include="Infrastructure\Influx\InfluxDbException.cs" />
+    <Compile Include="Infrastructure\Influx\InfluxDbResponse.cs" />
+    <Compile Include="Contracts\IRequestContent.cs" />
+    <Compile Include="Infrastructure\Json\JsonRequestContent.cs" />
+    <Compile Include="Enums\LogLevel.cs" />
     <Compile Include="Models\Point.cs" />
     <Compile Include="Models\QueryResult.cs" />
     <Compile Include="Models\WriteRequest.cs" />
-    <Compile Include="ObjectExtensions.cs" />
-    <Compile Include="HttpUtility.cs" />
-    <Compile Include="JsonSerializer.cs" />
+    <Compile Include="Helpers\ObjectExtensions.cs" />
+    <Compile Include="Helpers\HttpUtility.cs" />
+    <Compile Include="Infrastructure\Json\JsonSerializer.cs" />
     <Compile Include="InfluxDb.cs" />
     <Compile Include="Models\ContinuousQuery.cs" />
     <Compile Include="Models\Database.cs" />
@@ -72,8 +72,8 @@
     <Compile Include="Models\ShardSpace.cs" />
     <Compile Include="Models\User.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Stopwatch.cs" />
-    <Compile Include="TimeUnit.cs" />
+    <Compile Include="Helpers\Stopwatch.cs" />
+    <Compile Include="Enums\TimeUnit.cs" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="Microsoft.Threading.Tasks, Version=1.0.12.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
