+++ InfluxDB.Net/InfluxDb.cs
@@ -2,7 +2,13 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using InfluxDB.Net.Contracts;
+using InfluxDB.Net.Enums;
+using InfluxDB.Net.Helpers;
 using InfluxDB.Net.Infrastructure;
+using InfluxDB.Net.Infrastructure.Configuration;
+using InfluxDB.Net.Infrastructure.Influx;
+using InfluxDB.Net.Infrastructure.Validation;
 using InfluxDB.Net.Models;
 
 namespace InfluxDB.Net
