+++ InfluxDB.Net/Infrastructure/Configuration/DatabaseConfiguration.cs
@@ -2,7 +2,7 @@ using System;
 using System.Collections.Generic;
 using InfluxDB.Net.Models;
 
-namespace InfluxDB.Net
+namespace InfluxDB.Net.Infrastructure.Configuration
 {
     public class DatabaseConfiguration
     {
