+++ InfluxDB.Net/Infrastructure/Influx/InfluxDbException.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Net;
 
-namespace InfluxDB.Net
+namespace InfluxDB.Net.Infrastructure.Influx
 {
     public class InfluxDbException : Exception
     {
